requirements.txt
Django
asgiref
certifi
chardet
django-cors-headers
django-oauth-toolkit
djangorestframework
idna
amqp
oauthlib
pytz
requests
sqlparse
urllib3
elasticsearch-dsl>=7.0.0,<8.0.0
django-elasticsearch-dsl-drf==0.20.8
psycopg2-binary
celery==4.4.0
kombu>=4.6.10,<4.7
django-celery-results
django-modeltranslation
flake8
black
django-querysetsequence